Add basic extraction of propTypes into types that the compiler can check.
React.PropTypes.* expressions in the propTypes object literal are converted into
a record type that used as the type of the "props" member. Props are also
checked at element creation time.

Disabled by default for now, mixins are not handled, and there's no mechanism
for specifying the type of custom prop type checkers.

For #1