When using a minified build of React, add (and use) an alias for React.createElement.
Normally React.createElement calls are not renamed at all, due to React being
an extern and createElement showing up in the built-in browser DOM externs. By
adding an alias and then rewriting calls (see visitReactCreateElement) we allow
the compiler to rename the function used at all the calls. This is most
beneficial before gzip, but when after gzip there is still some benefit (due to
how many such calls there usually are).