Add option to generate a React externs SourceFile.
Useful when compiling React entirely separately.

To avoid code duplication, moves generation of the DOM factory functions and
event handler names to be done as string operations, instead of on the AST of
the types.js file. Makes things less verbose too.

Requires that the compiler input language be at least ES5, to allow the "var"
DOM factory function to be created.