Don't allow props of components annotated with @export to be renamed.
If the component is supposed to be accessible from non-compiled code, then it
should be possible to provide props to it too.

Done by injecting the prop names (from propTypes) into the list of externs.
This means that other properties with the same name won't get renamed, but that
is safer and simplier than trying to avoid renaming specifically within the
component.