Add an alias for React.createClass to use in minified builds
Along the lines of 5a372cfcfdf1a8ab6fb2f334eb8712141e463853, adds an
alias so that the symbol can be fully minified (before it was remaining
as React.[one letter variable]). The wins are nowhere near as
significant as those for createElement (which was much more frequent),
but it is a straightforward change.