Make type checking of prop types when using spread smarter
If a prop was required, we would require it to be present in all
object literal arguments to the spread call, which didn't work in
practice (because each might only contain a subset of props).

Generate another record type where all props are optional, and use
that when validating spreads (will stil catch mismatches where the prop
is present but has the wrong type).

CC=@jesseditson I think you ran into this sort of recently