Make sure we include parameters in mixin interface methods
For a mixin we generate an interface. We were not setting the parameters
in the methods. Like this

```js
class Mixin extends React.Component {
    /** @param {number} x */
    method(x) {}
}
ReactSupport.declareMixin(Mixin);
```

generated:

```js
/** @interface */
function MixinInterface() {}
/** @param {number} x */
MixinInterface.prototype.method = function() {}
```

The last line above needs to be:

```js
MixinInterface.prototype.method = function(x) {}
```

or `x` gets marked as optional which leads to issues when overriding.